requirements.txt
Flask
mysql-connector
click
flask-CLI
requests
Flask-WTF
python-dotenv